+++ packages/spacecat-shared-content-client/test/clients/content-client.test.js
@@ -16,8 +16,10 @@ import { expect, use } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import esmock from 'esmock';
 import sinon from 'sinon';
+import sinonChai from 'sinon-chai';
 
 use(chaiAsPromised);
+use(sinonChai);
 
 describe('ContentClient', () => {
   let sandbox;
@@ -37,11 +39,17 @@ describe('ContentClient', () => {
     getConfig: () => ({ content: { source: { type: 'onedrive' } } }),
   };
 
-  const sampleMetadata = new Map([
-    ['title', 'Test Page'],
-    ['description', 'Test description'],
-    ['keywords', 'test, metadata'],
-  ]);
+  const sampleMetadata = new Map(
+    [['title', { value: 'Test Page', type: 'text' }],
+      ['description', { value: 'Test description', type: 'text' }],
+      ['keywords', { value: 'test, metadata', type: 'text' }]],
+  );
+
+  const existingRedirects = [
+    { from: '/test-A', to: '/test-B' },
+    { from: '/test-C', to: '/test-D' },
+    { from: '/test-B', to: '/test-D' },
+  ];
 
   const createContentClient = async (getPageMetadata) => {
     const contentSDK = sinon.stub().returns({
@@ -59,6 +67,20 @@ describe('ContentClient', () => {
       getPageMetadata: getError
         ? sinon.stub().rejects(new Error(errorMessage)) : sinon.stub().resolves(new Map()),
       updatePageMetadata: sinon.stub().resolves(updateError ? { status: 500 } : { status: 200 }),
+      getRedirects: getError
+        ? sinon.stub().rejects(new Error(errorMessage)) : sinon.stub().resolves(existingRedirects),
+      appendRedirects: sinon.stub().resolves(updateError ? { status: 500 } : { status: 200 }),
+    });
+
+    return esmock('../../src/clients/content-client.js', {
+      '@adobe/spacecat-helix-content-sdk': { createFrom: contentSDK },
+    });
+  };
+
+  const createContentClientForRedirects = async (getRedirects) => {
+    const contentSDK = sinon.stub().returns({
+      getRedirects: sinon.stub().resolves(getRedirects),
+      appendRedirects: sinon.stub().resolves({ status: 200 }),
     });
 
     return esmock('../../src/clients/content-client.js', {
@@ -221,8 +243,8 @@ describe('ContentClient', () => {
     it('throws an error if raw client has non-200 status', async () => {
       ContentClient = await createErrorContentClient(false, true, 'Error updating page metadata');
       const metadata = new Map([
-        ['lang', 'en'],
-        ['keywords', 'test, metadata'],
+        ['lang', { value: 'en', type: 'text' }],
+        ['keywords', { value: 'test, metadata', type: 'text' }],
       ]);
       const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
       const path = '/test-path';
@@ -231,8 +253,8 @@ describe('ContentClient', () => {
 
     it('updates page metadata with valid metadata', async () => {
       const metadata = new Map([
-        ['lang', 'en'],
-        ['keywords', 'test, metadata'],
+        ['lang', { value: 'en', type: 'text' }],
+        ['keywords', { value: 'test, metadata', type: 'text' }],
       ]);
       const expectedMetadata = new Map([...sampleMetadata, ...metadata]);
 
@@ -278,19 +300,19 @@ describe('ContentClient', () => {
         ['description', ''], // Invalid value
       ]);
 
-      await expect(client.updatePageMetadata('/test-path', metadata)).to.be.rejectedWith('Metadata value for key description must be a string');
+      await expect(client.updatePageMetadata('/test-path', metadata)).to.be.rejectedWith('Metadata value for key description must be a object that has a value and type');
     });
 
     it('overwrites existing metadata by default when updating', async () => {
       const newMetadata = new Map([
-        ['description', 'New description'],
-        ['author', 'New Author'],
+        ['description', { value: 'New description', type: 'text' }],
+        ['author', { value: 'New Author', type: 'text' }],
       ]);
       const expectedMetadata = new Map([
-        ['title', 'Test Page'], // Original key remains
-        ['description', 'New description'], // Overwritten
-        ['author', 'New Author'], // Added
-        ['keywords', 'test, metadata'], // Original key remains
+        ['title', { value: 'Test Page', type: 'text' }], // Original key remains
+        ['description', { value: 'New description', type: 'text' }], // Overwritten
+        ['author', { value: 'New Author', type: 'text' }], // Added
+        ['keywords', { value: 'test, metadata', type: 'text' }], // Original key remains
       ]);
 
       ContentClient = await createContentClient(sampleMetadata);
@@ -305,14 +327,14 @@ describe('ContentClient', () => {
 
     it('merges without overwriting when overwrite option is false', async () => {
       const newMetadata = new Map([
-        ['description', 'New description'],
-        ['author', 'New Author'],
+        ['description', { value: 'New description', type: 'text' }],
+        ['author', { value: 'New Author', type: 'text' }],
       ]);
       const expectedMetadata = new Map([
-        ['description', 'Test description'], // Original key remains
-        ['keywords', 'test, metadata'], // Original key remains
-        ['title', 'Test Page'], // Original key remains
-        ['author', 'New Author'], // Added
+        ['description', { value: 'Test description', type: 'text' }], // Original key remains
+        ['keywords', { value: 'test, metadata', type: 'text' }], // Original key remains
+        ['title', { value: 'Test Page', type: 'text' }], // Original key remains
+        ['author', { value: 'New Author', type: 'text' }], // Added
       ]);
 
       ContentClient = await createContentClient(sampleMetadata);
@@ -325,6 +347,96 @@ describe('ContentClient', () => {
 
       expect(updatedMetadata).to.deep.equal(expectedMetadata);
       expect(client.rawClient.updatePageMetadata.calledOnceWith('/test-path', expectedMetadata)).to.be.true;
+    });
+  });
+
+  describe('updateRedirects', () => {
+    it('throws an error if raw client has non-200 status', async () => {
+      ContentClient = await createErrorContentClient(false, true, 'Error updating redirects');
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await expect(client.updateRedirects([{ from: '/A', to: '/B' }])).to.be.rejectedWith('Failed to update redirects');
+    });
+    it('throws an error if new redirects are not an array', async () => {
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await expect(client.updateRedirects({})).to.be.rejectedWith('Redirects must be an array');
+    });
+    it('throws an error if new redirects are an empty array', async () => {
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await expect(client.updateRedirects([])).to.be.rejectedWith('Redirects must not be empty');
+    });
+    it('throws an error if new redirects contains an invalid entry', async () => {
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await expect(client.updateRedirects([{ from: '/A', to: '/B' }, 'malformed-redirect'])).to.be.rejectedWith('Redirect must be an object');
+      await expect(client.updateRedirects([{ from: '/A', to: '/B' }, { from: '', to: '/B' }])).to.be.rejectedWith('Redirect must have a valid from path');
+      await expect(client.updateRedirects([{ from: '/A', to: '/B' }, { from: '/A', to: '' }])).to.be.rejectedWith('Redirect must have a valid to path');
+      await expect(client.updateRedirects([{ from: 'A', to: '/B' }, { from: '/A', to: '/C' }])).to.be.rejectedWith('Invalid redirect from path: A');
+      await expect(client.updateRedirects([{ from: '/A', to: 'B' }, { from: '/A', to: '/C' }])).to.be.rejectedWith('Invalid redirect to path: B');
+      await expect(client.updateRedirects([{ from: '/A', to: '/B' }, { from: '/A', to: '/A' }])).to.be.rejectedWith('Redirect from and to paths must be different');
+    });
+    it('update success', async () => {
+      const newRedirects = [
+        { from: '/test-X', to: '/test-Y' },
+      ];
+
+      ContentClient = await createContentClientForRedirects(existingRedirects);
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await client.updateRedirects(newRedirects);
+      await expect(client.rawClient.appendRedirects.calledOnceWith(sinon.match([{ from: '/test-X', to: '/test-Y' }]))).to.be.true;
+    });
+    it('update success ignores duplicates', async () => {
+      const newRedirects = [
+        { from: '/test-A', to: '/test-B' },
+        { from: '/test-X', to: '/test-Y' },
+        { from: '/test-B', to: '/test-D' },
+        { from: '/test-X', to: '/test-Y' },
+      ];
+
+      ContentClient = await createContentClientForRedirects(existingRedirects);
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await client.updateRedirects(newRedirects);
+      await expect(client.rawClient.appendRedirects.calledOnceWith(sinon.match([{ from: '/test-X', to: '/test-Y' }]))).to.be.true;
+    });
+    it('detect cycles in new redirects', async () => {
+      const newRedirects = [
+        { from: '/test-D', to: '/test-A' },
+        { from: '/test-C', to: '/test-E' },
+      ];
+      ContentClient = await createContentClientForRedirects(existingRedirects);
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await client.updateRedirects(newRedirects);
+      await expect(client.rawClient.appendRedirects.calledOnceWith(sinon.match([{ from: '/test-C', to: '/test-E' }]))).to.be.true;
+    });
+    it('detect cycles in current redirects', async () => {
+      const newRedirects = [
+        { from: '/test-I', to: '/test-J' },
+      ];
+      const cycleRedirects = [
+        { from: '/test-A', to: '/test-C' },
+        { from: '/test-C', to: '/test-E' },
+        { from: '/test-E', to: '/test-A' },
+      ];
+      ContentClient = await createContentClientForRedirects(cycleRedirects);
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await expect(client.updateRedirects(newRedirects)).to.be.rejectedWith('Redirect cycle detected');
+    });
+    it('does not call rawClient when all redirects are duplicates', async () => {
+      const newRedirects = [
+        { from: '/test-A', to: '/test-B' },
+        { from: '/test-C', to: '/test-D' },
+      ];
+      ContentClient = await createContentClientForRedirects(existingRedirects);
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await client.updateRedirects(newRedirects);
+      await expect(client.rawClient.appendRedirects).to.not.have.been.called;
+    });
+    it('does not call rawClient when there are no valid redirects', async () => {
+      const newRedirects = [
+        { from: '/test-D', to: '/test-A' },
+      ];
+      ContentClient = await createContentClientForRedirects(existingRedirects);
+      const client = ContentClient.createFrom(context, siteConfigGoogleDrive);
+      await client.updateRedirects(newRedirects);
+      await expect(client.rawClient.appendRedirects).to.not.have.been.called;
     });
   });
 });